pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Yaroslav <benyros@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.73.0"
SQLAlchemy = "^1.4.31"
uvicorn = "^0.17.4"
python-multipart = "^0.0.5"
alembic = "^1.7.6"
python-dotenv = "^0.19.2"
databases = "^0.5.5"
sqlmodel = "^0.0.6"
asyncpg = "^0.25.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.931"
sqlalchemy-stubs = "^0.4"
flake8 = "^4.0.1"
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
